add input_check

+++ push_swap/srcs/libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: minkyuki <minkyuki@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 16:53:42 by minkyuki          #+#    #+#             */
-/*   Updated: 2022/11/09 09:06:13 by minkyuki         ###   ########.fr       */
+/*   Updated: 2022/11/28 18:40:54 by minkyuki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int			sign;
 	int			i;
@@ -37,5 +37,5 @@ int	ft_atoi(const char *str)
 		i++;
 	}
 	result *= sign;
-	return ((int)result);
+	return (result);
 }
